% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bgee.R
\docType{class}
\name{Bgee-class}
\alias{Bgee}
\alias{Bgee-class}
\title{Retrieving the Bgee database data}
\value{
\itemize{
 \item{\code{get_annotation()} returns a list of the annotation of experiments for chosen species.}
 \item{\code{get_data()}, if experiment ID is empty, returns a list of experiments. If specified experiment ID, then returns the dataframe of the chosen experiment}
 \item{\code{format_data()}, if experiment ID is empty, returns a list of ExpressionSet objects. If specified experiment ID, then returns an ExpressionSet object}}
}
\description{
A Reference Class to give annotation available on Bgee for particular species and the requested data (rna_seq, affymetrix)
}
\details{
The expression calls come from Bgee (http://r.bgee.org), that integrates different expression data types (RNA-seq, Affymetrix microarray, ESTs, or in-situ hybridizations) in multiple animal species. Expression patterns are based exclusively on curated "normal", healthy, expression data (e.g., no gene knock-out, no treatment, no disease), to provide a reference of normal gene expression.
This Class retrieves annotation of all experiments in Bgee database (get_annotation), downloading the data (get_data), and formating the data into expression matrix (format_data). See examples and vignette.
}
\section{Fields}{

\describe{
\item{\code{species}}{A character of species name as listed from Bgee. The species are:
\itemize{
   \item{"Anolis_carolinensis"}
   \item{"Bos_taurus"}
   \item{"Caenorhabditis_elegans"}
   \item{"Danio_rerio"}
   \item{"Drosophila_melanogaster"}
   \item{"Gallus_gallus"}
   \item{"Gorilla_gorilla"}
   \item{"Homo_sapiens"}
   \item{"Macaca_mulatta"}
   \item{"Monodelphis_domestica"}
   \item{"Mus_musculus"}
   \item{"Ornithorhynchus_anatinus"}
   \item{"Pan_paniscus"}
   \item{"Pan_troglodytes"}
   \item{"Rattus_norvegicus"}
   \item{"Sus_scrofa"}
   \item{"Xenopus_tropicalis"}}
Homo sapiens is the default species.}

\item{\code{datatype}}{A character of data platform. Two types of datasets can be downloaded:
\itemize{
     \item{"rna_seq"}
     \item{"affymetrix"}}
By default, RNA-seq data is retrieved.}

\item{\code{experiment.id}}{An ArrayExpress or GEO accession, e.g., GSE30617
On default is NULL: takes all available experiments for specified species and datatype.}

\item{\code{data}}{A dataframe of downloaded Bgee data.}

\item{\code{calltype}}{A character.
 \itemize{
   \item{"expressed"}
   \item{"expressed high quality"}
   \item{"all"}}
Retrieve intensities only for expressed (present) genes, expressed high quality genes, or all genes. The default is expressed.}

\item{\code{stats}}{A character. The expression values can be retrieved in RPKMs and raw counts:
 \itemize{
   \item{"rpkm"}
   \item{"counts"}
   \item{"intensities"}
   }
The default is RPKMs for RNA-seq and intensities for microarray.}
}}
\examples{
{
 bgee <- Bgee$new(species = "Mus_musculus", datatype = "rna_seq")
 annotation_bgee_mouse <- bgee$get_annotation()
 data_bgee_mouse <- bgee$get_data()
 data_bgee_mouse_gse30617 <- bgee$get_data(experiment.id = "GSE30617")
 gene.expression.mouse.rpkm <- bgee$format_data(data_bgee_mouse_gse30617,
 calltype = "expressed", stats = "rpkm")
 }


}
\author{
Andrea Komljenovic \email{andrea.komljenovic at unil.ch}.
}

