% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteLocalData.R
\name{deleteLocalData}
\alias{deleteLocalData}
\title{Delete local database for the species of the reference class Bgee object}
\usage{
deleteLocalData(myBgeeObject)
}
\arguments{
\item{myBgeeObject}{A Reference Class Bgee object, notably specifying the targeted species}
}
\description{
This function delete all data present in the local database of one species.
}
\examples{
{
  bgee <- Bgee$new(species = "Mus_musculus", dataType = "rna_seq")
  deleteLocalData(bgee)
}

}
\author{
Julien Wollbrett
}
