% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_wrapper.R
\name{downloadCellProcessedData_galaxy}
\alias{downloadCellProcessedData_galaxy}
\title{Retrieve Bgee processed expression values at cell level for single cell data.}
\usage{
downloadCellProcessedData_galaxy(
  species = NULL,
  experimentId = NULL,
  cellProcessedDataFile = NULL
)
}
\arguments{
\item{species}{name of the species for which annotation have to be be retrieved (e.g Homo_sapiens)}

\item{experimentId}{Filter allowing to specify one ArrayExpress or GEO accession, e.g.,
GSE43721. Can not be null}

\item{cellProcessedDataFile}{Name of the file where processed expression values at cell
level is stored}
}
\description{
This function download the processed gene count sparse matrix of single cell experiments
in Bgee. These data are available for all droplet based single-cell RNA-seq and full length
single-cell RNA-seq. These files are stored in the H5AD format.
}
\author{
Julien Wollbrett
}
