% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_wrapper.R
\name{getAnnotation_galaxy}
\alias{getAnnotation_galaxy}
\title{Galaxy adapter to the \code{getAnnotation} function}
\usage{
getAnnotation_galaxy(
  species = NULL,
  dataType = character(0),
  experimentAnnotationPath = NULL,
  sampleAnnotationPath = NULL
)
}
\arguments{
\item{species}{name of the species for which annotation have to be be retrieved (e.g Homo_sapiens)}

\item{dataType}{The datatype for which annotation have to be retrieved. To be chosen among:
\itemize{
\item{"rna_seq: target only bulk RNA-seq data"}
\item{"sc_full_length: target only full length single-cell RNA-seq data"}
\item{"sc_droplet_based: target only droplet based single-cell RNA-seq data"}
\item{"affymetrix: target only Affymtrix microarray data"}
}}

\item{experimentAnnotationPath}{path to the file containing experiment annotation}

\item{sampleAnnotationPath}{path to the file containing sample annotation}
}
\description{
Adapter of the \code{getAnnotation} function allowing to run it on Galaxy.
}
