% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBgeeRelease.R
\name{listBgeeRelease}
\alias{listBgeeRelease}
\title{List Bgee releases available to use with BgeeDB package}
\usage{
listBgeeRelease(release = NULL)
}
\arguments{
\item{release}{A character specifying a targeted release number. In the form "Release.subrelease" or "Release_subrelease", e.g., "13.2" or 13_2". If not specified, all available releases are shown.}
}
\value{
A data frame with information on Bgee releases.
}
\description{
Returns information on available Bgee releases, the access URL for FTP and webservice, and the date of release
}
\examples{
{
 listBgeeRelease()
}

}
\author{
Julien Roux, Julien Wollbrett
}

