% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBgeeSpecies.R
\name{listBgeeSpecies}
\alias{listBgeeSpecies}
\title{List of species in Bgee database}
\usage{
listBgeeSpecies(...)
}
\arguments{
\item{...}{an empty parameter}
}
\value{
A list of available RNA-seq and Affymetrix datasets
}
\description{
Takes a URL and returns a list of available genomes for different platforms in Bgee.
}

