% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrthologs.R, R/retrieveExpression.R
\name{listOrthologs}
\alias{listOrthologs}
\title{Retrieve orthologs from Bgee}
\usage{
listOrthologs(
  bgeeRelease = "current",
  downloadPath = getwd(),
  mandatorySpecies = NULL,
  optionalSpecies = NULL,
  referenceSpecies = NULL,
  onlyOneToOne = FALSE,
  onlyOneToMany = FALSE,
  removeDownloadFiles = FALSE,
  createFile = FALSE
)

listOrthologs(
  bgeeRelease = "current",
  downloadPath = getwd(),
  mandatorySpecies = NULL,
  optionalSpecies = NULL,
  referenceSpecies = NULL,
  onlyOneToOne = FALSE,
  onlyOneToMany = FALSE,
  removeDownloadFiles = FALSE,
  createFile = FALSE
)
}
\arguments{
\item{bgeeRelease}{(default current)}

\item{downloadPath}{path to the directory where orthologs will be downloaded}

\item{mandatorySpecies}{species for which it is mandatory to find orthologs.}

\item{optionalSpecies}{species for which we want to add orthologs if available. Will still retrieve
the orthologous genes from mandatory species even if no ortholog in these optionalSpecies}

\item{referenceSpecies}{Correspond to one species ID. Has to be provided only if onlyOneToOne is FALSE.
It will then retrieve all ortholog genes having more than one copy in mandatory and optional species.}

\item{onlyOneToOne}{boolean defining whether only one-to-one orthologs have to be retrieved (TRUE) or not (default : FALSE)}

\item{onlyOneToMany}{boolean defining whether only one-to-many orthologs have to be retrieved (TRUE) or not (default : FALSE)}

\item{removeDownloadFiles}{Does downloaded files have to be deleted once orthologous genes have been retrieved.
Do not delete files by default.}

\item{createFile}{By default the function return the data.frame of all orthologous genes. If \code{createFile} is TRUE then the
orthologous genes are saved in a file and no data is returned from this function}
}
\description{
This function allows to load all ortholog genes based on criteria provided
as arguments of this function. It is for instance possible to retrieve only one-to-one orthologs
or one-to-many orthologs. Please read the description of the arguments to properly tune the orthologs
to retrieve

This function allows to load all ortholog genes based on criteria provided
as arguments of this function. It is for instance possible to retrieve only one-to-one orthologs
or one-to-many orthologs. Please read the description of the arguments to properly tune the orthologs
to retrieve
}
\examples{
{
  listOrthologs(mandatorySpecies = c(9606,10090,7955,7227,13616), onlyOneToOne = TRUE)
}

{
  # 1-to-1 orthologs
  orthologs <- listOrthologs(mandatorySpecies = c(9606,10090,7955,7227,13616), onlyOneToOne = TRUE)
  # 1-to-1 orthologs saved in a file
  listOrthologs(mandatorySpecies = c(9606,10090,7955,7227,13616), onlyOneToOne = TRUE, createFile = TRUE)
  # all orthologs (1-to-1 and 1-to-many) from 10090, 9606 and 13616 with potential duplication in 9606 or 13616
  orthologs <- listOrthologs(referenceSpecies = 10090, mandatorySpecies = c(9606,13616))
  # 1-to-many orthologs from 10090, 9606 and 13616 with duplication in 9606 and 13616
  orthologs <- listOrthologs(referenceSpecies = 10090, mandatorySpecies = c(9606,13616), onlyOneToMany = TRUE)
}

}
\author{
Julien Wollbrett
}
