% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveExpression.R
\name{retrieveOrthologsExpression}
\alias{retrieveOrthologsExpression}
\title{Retrieve expression of 1-to-1 orthologous genes from different species but one single experiment using BgeeDB R package.}
\usage{
retrieveOrthologsExpression(
  bgeeRelease = "current",
  downloadPath = getwd(),
  experimentId = NULL,
  speciesIds = NULL,
  anatEntityIds = NULL,
  removeDownloadFiles = FALSE,
  onlyOneToOneOrthologs = FALSE,
  onlyOneToManyOrthologs = FALSE
)
}
\arguments{
\item{bgeeRelease}{(default current)}

\item{downloadPath}{path to the directory where expression data will be downloaded}

\item{experimentId}{the ID of the experiment for which bulk RNA-Seq expression data have to be retrieved.
Can not be NULL. The experiment ID needs to exist and to have libraries in Bgee for provided \code{speciesIds}}

\item{speciesIds}{species IDs for which we want to retrieve bulk RNA-Seq expression data. Can not be NULL.
The \code{speciesIds} needs to exist and to have expression data in Bgee for the specified \code{experimentId}.
If 1-to-many orthologs have to be retrieved then the reference species has to be provided as \code{referenceSpecies} and should
not be added to the list of \code{speciesIds}}

\item{anatEntityIds}{List of anatomical entity IDs for which expression data have to be retrieved. If NULL then all
anatomical entities will be retrieved.}

\item{removeDownloadFiles}{Does downloaded files have to be deleted once orthologous genes have been retrieved.
Do not delete files by default.}

\item{onlyOneToOneOrthologs}{boolean defining whether only one-to-one orthologs have to be retrieved (TRUE) or not (default : FALSE)}

\item{onlyOneToManyOrthologs}{boolean defining whether only one-to-many orthologs have to be retrieved (TRUE) or not (default : FALSE)}

\item{referenceSpecies}{Correspond to one species ID. Has to be provided only if onlyOneToOne is FALSE. It will then retrieve all ortholog
genes having more than one copy in mandatory and optional species.}
}
\description{
This function allows to load bulk RNA-Seq processed expression values from one experiment with libraries coming
from different species. Each row corresponds to a group of 1-to-1 orthologs and each column corresponds to one organ from one species.
}
\examples{
{
  retrieveOrthologsExpression(experimentId = "GSE30352", speciesIds = c(9606,10090, 13616), onlyOneToOneOrthologs = TRUE)
}

}
\author{
Julien Wollbrett
}
