% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topAnat.R
\name{topAnat}
\alias{topAnat}
\title{Load function allowing to perform GO-like enrichment of anatomical terms using the topGO package}
\usage{
topAnat(topAnatData)
}
\arguments{
\item{...}{}
}
\value{
...
}
\description{
This function loads the functions necessary to use topGO on the Uberon ontology instead of the GO ontology
}
\details{
To perform the enrichment test for expression in anatomical structures for each term of Uberon ontology (browsable at \url{http://www.ontobee.org/ontology/UBERON}), it is interesting to use the topGO package since it allows to propagate the mapping of gene to terms to parent terms, and it possesses a pannel of enrichment test and decorrelation methods
}
\examples{
  topAnat()

}
\author{
Julien Roux \email{julien.roux@unil.ch}.
}

