% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_wrapper.R
\name{topAnat_galaxy}
\alias{topAnat_galaxy}
\title{Run a GO-like enrichment of anatomical terms, mapped to genes by expression patterns}
\usage{
topAnat_galaxy(
  species = NULL,
  dataTypes = character(0),
  stageId = NULL,
  foregroundGenes = NULL,
  backgroundGenes = NULL,
  foregroundGenesFile = NULL,
  backgroundGenesFile = NULL,
  algorithm = "classic",
  statistics = "fisher",
  resultFile = NULL,
  nodeSize = 10,
  confidence = "silver"
)
}
\arguments{
\item{species}{name of the species for which annotation have to be be retrieved (e.g Homo_sapiens)}

\item{dataTypes}{The datatypes for which calls will to be retrieved. To be chosen among:
\itemize{
\item{"rna_seq: target only bulk RNA-seq data"}
\item{"sc_full_length: target only full length single-cell RNA-seq data"}
\item{"sc_droplet_based: target only droplet based single-cell RNA-seq data"}
\item{"affymetrix: target only Affymtrix microarray data"}
}}

\item{foregroundGenes}{the list of genes for which TopAnat will find anatomical entities that
have over or under-represented expression using annotations for that gene set, compared to the
background genes. Can be NULL if foregroundGenesFile is provided.}

\item{backgroundGenes}{the list of genes you want to consider as the universe in your analysis.
Optional. If NULL and backgroundGenesFile is also NULL, all genes in the species will be used as background.}

\item{foregroundGenesFile}{path to a one-column TSV file without header containing foreground genes.
Can be NULL if foregroundGenes is provided.}

\item{backgroundGenesFile}{path to a one-column TSV file without header containing background genes.
Optional. If NULL and backgroundGenes is also NULL, all genes in the species will be used as background.}

\item{algorithm}{Decorrelation algorithm used to take into account the topology of the anatomical
ontology (default = weight). By default \code{classic} is used. The full list of algorithm can be
retrieved with topGO::whichAlgorithms()}

\item{statistics}{Character string specifing which test to use. By default \code{fisher}is used. The full list
of tests can be retrieved with topGO::whichTests()}

\item{resultFile}{The file where GO-like enrichment of anatomical terms is stored}

\item{nodeSize}{Minimum number of genes mapped to a node for it to be tested. Default is 10.}

\item{confidence}{A character indicating if only high quality present calls should be
retrieved. For Bgee releases prior to 14, options are "all" (default) or "high_quality".
For Bgee release 14 and above, options are "silver" (default) and "gold".}
}
\description{
This function GO-like enrichment of anatomical terms using the Uberon ontology.
}
\examples{
{
foregroundGenes <- c("ENSBTAG00000000011","ENSBTAG00000000014","ENSBTAG00000000016",
"ENSBTAG00000000026","ENSBTAG00000000039","ENSBTAG00000000040",
"ENSBTAG00000000042","ENSBTAG00000000050","ENSBTAG00000000056",
"ENSBTAG00000000064","ENSBTAG00000000067","ENSBTAG00000000071",
"ENSBTAG00000000072","ENSBTAG00000000080","ENSBTAG00000000081")
backgroundGenes <- c("ENSBTAG00000000011","ENSBTAG00000000014","ENSBTAG00000000016",
"ENSBTAG00000000026","ENSBTAG00000000039","ENSBTAG00000000040",
"ENSBTAG00000000042","ENSBTAG00000000050","ENSBTAG00000000056",
"ENSBTAG00000000064","ENSBTAG00000000067","ENSBTAG00000000071",
"ENSBTAG00000000072","ENSBTAG00000000080","ENSBTAG00000000081",
"ENSBTAG00000000084","ENSBTAG00000000091","ENSBTAG00000000099",
"ENSBTAG00000000111","ENSBTAG00000000123","ENSBTAG00000000132",
"ENSBTAG00000000153","ENSBTAG00000000162","ENSBTAG00000000163",
"ENSBTAG00000000169","ENSBTAG00000000179","ENSBTAG00000000197",
"ENSBTAG00000000199","ENSBTAG00000000202","ENSBTAG00000000203",
"ENSBTAG00000000204","ENSBTAG00000000213","ENSBTAG00000000215",
"ENSBTAG00000000223","ENSBTAG00000000224","ENSBTAG00000000225",
"ENSBTAG00000000236","ENSBTAG00000000250","ENSBTAG00000000251",
"ENSBTAG00000000252","ENSBTAG00000000253","ENSBTAG00000000261",
"ENSBTAG00000000274","ENSBTAG00000000277","ENSBTAG00000000279",
"ENSBTAG00000000285","ENSBTAG00000000286","ENSBTAG00000000287",
"ENSBTAG00000000289","ENSBTAG00000000297","ENSBTAG00000000305",
"ENSBTAG00000000312","ENSBTAG00000000328","ENSBTAG00000000335",
"ENSBTAG00000000341","ENSBTAG00000000343","ENSBTAG00000000354",
"ENSBTAG00000000355","ENSBTAG00000000356","ENSBTAG00000000365",
"ENSBTAG00000000372","ENSBTAG00000000379","ENSBTAG00000000380",
"ENSBTAG00000000382","ENSBTAG00000000396","ENSBTAG00000000404",
"ENSBTAG00000000405","ENSBTAG00000000406","ENSBTAG00000000411",
"ENSBTAG00000000425","ENSBTAG00000000434","ENSBTAG00000000435",
"ENSBTAG00000000438","ENSBTAG00000000448","ENSBTAG00000000451",
"ENSBTAG00000000454","ENSBTAG00000000456","ENSBTAG00000000457",
"ENSBTAG00000000459","ENSBTAG00000000462","ENSBTAG00000000469",
"ENSBTAG00000000470","ENSBTAG00000000484","ENSBTAG00000000497",
"ENSBTAG00000000501","ENSBTAG00000009707","ENSBTAG00000026266",
"ENSBTAG00000021992","ENSBTAG00000005353","ENSBTAG00000005333",
"ENSBTAG00000006424","ENSBTAG00000026972","ENSBTAG00000010799",
"ENSBTAG00000014614","ENSBTAG00000045757","ENSBTAG00000046332",
"ENSBTAG00000008394")
topAnat_galaxy(species = "Bos_taurus", stageId = "UBERON:0000092",
              foregroundGenes = foregroundGenes, backgroundGenes = backgroundGenes,
              resultFile = "result.tsv")
}

}
\author{
Julien Wollbrett
}
