% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bgee.R
\docType{class}
\name{Bgee-class}
\alias{Bgee}
\alias{Bgee-class}
\title{This function gives annotation available on Bgee for particular species and the requested data (rna_seq, affymetrix)}
\arguments{
\item{species}{a character of species name as listed from Bgee.
Options are:          
         "Anolis_carolinensis"      
         "Bos_taurus"               
         "Caenorhabditis_elegans"
         "Danio_rerio" 
         "Drosophila_melanogaster"  
         "Gallus_gallus"            
         "Gorilla_gorilla"          
         "Homo_sapiens"
         "Macaca_mulatta"           
         "Monodelphis_domestica"    
         "Mus_musculus"
         "Ornithorhynchus_anatinus" 
         "Pan_paniscus"             
         "Pan_troglodytes"          
         "Rattus_norvegicus"        
         "Sus_scrofa"               
         "Xenopus_tropicalis"}

\item{datatype}{a character of data platform.
Options are: 
         "rna_seq"
         "affymetrix"}

\item{experiment.id}{a character. 
On default is NULL: takes all available data for that species
GSE[0-9]+: takes specified experiment, eg. GSE30617}
}
\value{
tables of metadata or expression data
get_annotation() - tables of experiments and libraries, saved in list
get_data() - if all data, saved in the directory and as a list in R; if a specific experiment, saved in directiory and only as a table in R
}
\description{
This function gives annotation available on Bgee for particular species and the requested data (rna_seq, affymetrix)
}
\examples{


# note that the vignette shows more examples
bgee <- Bgee$new(species = "Mus_musculus", datatype = "rna_seq")
annotation_bgee_mouse <- bgee$get_annotation()
data_bgee_mouse <- bgee$get_data()
# download RPKMs and counts only for GSE30617 for Mus musculus
data_bgee_mouse_gse30617 <- bgee$get_data(experiment.id = "GSE30617")

}

