% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bgee.R
\docType{class}
\name{Bgee-class}
\alias{Bgee}
\alias{Bgee-class}
\title{Retrieving the Bgee database data}
\value{
\itemize{
 \item{\code{get_annotation()} returns a list of the annotation of experiments for chosen species.}
 \item{\code{get_data()}, if experiment ID is empty, returns a list of experiments. If specified experiment ID, then returns the dataframe of the chosen experiment}
 \item{\code{format_data()}, if experiment ID is empty, returns a list of ExpressionSet objects. If specified experiment ID, then returns an ExpressionSet object}}
}
\description{
A Reference Class to give annotation available on Bgee for particular species and the requested data (rna_seq, affymetrix)
}
\details{
The expression data come from Bgee (http://.bgee.org), that integrates different expression data types (RNA-seq, Affymetrix microarray, ESTs, or in-situ hybridizations) in multiple animal species. Expression patterns are based exclusively on curated "normal", healthy, expression data (e.g., no gene knock-out, no treatment, no disease), to provide a reference of normal gene expression.
This Class retrieves annotation of experiments in Bgee database (get_annotation), downloads the data (get_data), and formats the data into expressionSet object (format_data). See examples and vignette.
}
\section{Fields}{

\describe{
\item{\code{species}}{A character indicating the species to be used, in the form "Genus_species", or a numeric indicating the species NCBI taxonomic id. Only species with quantitative expression data in Bgee will work (RNA-seq and Affymetrix microarray data). See the listBgeeSpecies() function to get the list of species available in the Bgee release used for these two data types.}

\item{\code{datatype}}{A character of data platform. Quantitative expression levels can be downloaded for two data types:
\itemize{
     \item{"rna_seq"}
     \item{"affymetrix"}}}

\item{\code{experiment.id}}{An ArrayExpress or GEO accession, e.g., GSE30617. Used in get_data() function.
On default is NULL: takes all available experiments for specified species and datatype.}

\item{\code{pathToData}}{Path to the directory where the data files are stored. By default the working directory is used.}

\item{\code{release}}{Bgee release number to download data from, in the form "Release.subrelease" or "Release_subrelease", e.g., "13.2" or 13_2". Will work for release >=13.2. By default, the latest relase of Bgee is used.}

\item{\code{calltype}}{A character.
 \itemize{
   \item{"present"}
   \item{"present high quality"}
   \item{"all"}}
Retrieve intensities only for present (expressed) genes, present high quality genes, or all genes. The default is present.}

\item{\code{stats}}{A character indicating what expression values should be used in the formatted data expressionSet object.
\itemize{
  \item{"rpkm" for RNA-seq}
  \item{"counts" for RNA-seq}
  \item{"tpm" for RNA-seq (Bgee release 14 and above only)}
  \item{"intensities" for Affymetrix microarrays}
  }}
}}
\examples{
{
 bgee <- Bgee$new(species = "Mus_musculus", datatype = "rna_seq")
 annotation_bgee_mouse <- bgee$get_annotation()
 data_bgee_mouse <- bgee$get_data()
 data_bgee_mouse_gse30617 <- bgee$get_data(experiment.id = "GSE30617")
 gene.expression.mouse.rpkm <- bgee$format_data(data_bgee_mouse_gse30617,
 calltype = "present", stats = "rpkm")
 }


}
\author{
Andrea Komljenovic and Julien Roux.
}

