% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bgee.R
\docType{class}
\name{Bgee-class}
\alias{Bgee}
\alias{Bgee-class}
\title{Retrieving the Bgee database data}
\description{
A Reference Class to give annotation available on Bgee for particular species and the requested data (rna_seq, affymetrix)
}
\section{Fields}{

\describe{
\item{\code{species}}{A character of species name as listed from Bgee.
Options are:
         "Anolis_carolinensis",
         "Bos_taurus",
         "Caenorhabditis_elegans",
         "Danio_rerio",
         "Drosophila_melanogaster",
         "Gallus_gallus",
         "Gorilla_gorilla",
         "Homo_sapiens",
         "Macaca_mulatta",
         "Monodelphis_domestica",
         "Mus_musculus",
         "Ornithorhynchus_anatinus",
         "Pan_paniscus",
         "Pan_troglodytes",
         "Rattus_norvegicus",
         "Sus_scrofa",
         "Xenopus_tropicalis"}

\item{\code{datatype}}{A character of data platform.
Options are:
         "rna_seq",
         "affymetrix"}

\item{\code{experiment.id}}{A character.
On default is NULL: takes all available data for that species.
If GSE[0-9]+: takes specified experiment, eg. GSE30617.}

\item{\code{data}}{A dataframe of downloaded Bgee data.}

\item{\code{calls}}{A character.
Options are: "present" or "absent"}

\item{\code{stats}}{A character.
Options are: "rpkm" , "counts", "tpm"}
}}
\examples{
\dontrun{
bgee <- Bgee$new(species = "Mus_musculus", datatype = "rna_seq")
annotation_bgee_mouse <- bgee$get_annotation()
data_bgee_mouse <- bgee$get_data()
data_bgee_mouse_gse30617 <- bgee$get_data(experiment.id = "GSE30617")
}


}
\author{
Andrea Komljenovic \email{andrea.komljenovic at unil.ch}.
}

