% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Retrieve Bgee RNA-seq or Affymetrix data.}
\usage{
getData(myBgeeObject, experimentId = NULL)
}
\arguments{
\item{myBgeeObject}{A Reference Class Bgee object, notably specifying the targeted species and data type.}

\item{experimentId}{An ArrayExpress or GEO accession, e.g., GSE43721. Default is NULL: takes all available experiments for targeted species and data type.}
}
\value{
If experimentId is not specified, returns a list of data frames with data from all experiments for targeted species and data type. If experimentId is specified, returns a data frame with data from this experiment.
}
\description{
This function loads the quantitative expression data and presence calls for samples available from Bgee (rna_seq, affymetrix).
}
\examples{
{
  bgee <- Bgee$new(species = "Mus_musculus", dataType = "rna_seq")
  dataMouse <- getData(bgee)
  dataMouseGSE43721 <- getData(bgee, experimentId = "GSE43721")
}

}
\author{
Andrea Komljenovic and Julien Roux.
}

