% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBgeeSpecies.R
\name{listBgeeSpecies}
\alias{listBgeeSpecies}
\title{List species with Affymetrix or RNA-seq data in the Bgee database}
\usage{
listBgeeSpecies(...)
}
\arguments{
\item{...}{an empty parameter}
}
\value{
A list of species with available Affymetrix or RNA-seq data in the Bgee
}
\description{
Returns a list of available genomes for different platforms in Bgee.
}
\author{
Andrea Komljenovic \email{andrea.komljenovic at unil.ch}.
}

