% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTable.R
\name{makeTable}
\alias{makeTable}
\title{Formats the results obtained from gene set enrichment test on anatomical structures}
\usage{
makeTable(topAnatData, topAnatObject, results, cutoff)
}
\arguments{
\item{topAnatData}{A list produced by the function loadTopAnatData().}

\item{topAnatObject}{An object produced by the function topAnat().}

\item{results}{A result object, produced for example by the runtest() function of topGO}

\item{cutoff}{An FDR cutoff between 0 and 1. Only terms with FDR lower than this cutoff will be in the output. Default is 1.}
}
\value{
A data frame with significantly enriched anatomical structures, sorted by FDR
}
\description{
This function loads the results from the test function (topGO package) and creates an output table with organ names, fold enrichment and FDR
}
\examples{
  myTopAnatData <- loadTopAnatData(species = "Mus_musculus", datatype = "rna_seq")
  geneList <- c(0,1,0,1)
  names(geneList) <- c("gene1","gene3","gene3","gene4")
  myTopAnatObject <- topAnat(myTopAnatData, geneList)
  TO DO...

}
\author{
Julien Roux \email{julien.roux@unil.ch}.
}

