% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topAnat.R
\name{topAnat}
\alias{topAnat}
\title{Produces an object allowing to perform GO-like enrichment of anatomical terms using the topGO package}
\usage{
topAnat(topAnatData, geneList, nodeSize = 10, ...)
}
\arguments{
\item{topAnatData}{a list produced by the function loadTopAnatData().}

\item{geneList}{Vector indicating foreground and background genes. Names of the vector indicate teh background genes. Vector values are 0 (gene not in foreground) and 1 (gene in foreground).}

\item{nodeSize}{Minimum number of genes mapped to a node for it to be tested. Default is 10.}

\item{...}{Additional parameters as passed to topGOdata object of topGO package.}
}
\value{
topAnatObject, a topGO-compatibe object ready for gene set enrichment testing.
}
\description{
This function loads the functions necessary to use topGO on the Uberon ontology instead of the GO ontology. It then produces a topAnatObject, ready to use for gene set enrichment testing using functions from the topGO package.
}
\details{
To perform the enrichment test for expression in anatomical structures for each term of Uberon ontology (browsable at \url{http://www.ontobee.org/ontology/UBERON}), it is interesting to use the topGO package since it allows to propagate the mapping of gene to terms to parent terms, and it possesses a pannel of enrichment tests and decorrelation methods.
}
\examples{
 \dontrun{
  myTopAnatData <- loadTopAnatData(species = "Mus_musculus", datatype = "rna_seq")
  geneList <- c(0,1,0,1)
  names(geneList) <- c("gene1","gene3","gene3","gene4")
  myTopAnatObject <- topAnat(myTopAnatData, geneList)
}

}
\author{
Julien Roux \email{julien.roux@unil.ch}.
}

